refactor(cli): split cli.py by command group
Move CLI command bodies and batch strategies out of the 807-LOC cli.py
into src/imagecli/commands/. cli.py becomes a 24-LOC thin Typer wiring.

Layout:
- commands/_helpers.py: resolve_face_image(s), load_config, run_generate
- commands/_batch_common.py: Progress bar, peak VRAM, param resolution
- commands/_batch_{all_on_gpu,two_phase,sequential}.py: one strategy each
- commands/{generate,batch,meta,serve}.py: command modules, each exposes
  register(app) so cli.py owns the single Typer instance

Drop the cli.py exemption from tools/file_exemptions.txt.
Tests patch at new locations; add unit tests for resolve_face_image(s).

Closes #54
